feat: generate customer links html
